+++ csfunctions/events/document_release_check.py
@@ -5,13 +5,13 @@ from pydantic import BaseModel, Field
 from csfunctions.objects import Document, Part
 
 from .base import BaseEvent, EventNames
-from .dialog_data import DocumentReleaseDialogData
+from .dialog_data import DocumentReleasedDialogData
 
 
 class DocumentReleaseCheckData(BaseModel):
     documents: list[Document] = Field(..., description="List of documents that will be released.")
     attached_parts: list[Part] = Field(..., description="List of parts that belong to the documents")
-    dialog_data: DocumentReleaseDialogData
+    dialog_data: DocumentReleasedDialogData
 
 
 class DocumentReleaseCheckEvent(BaseEvent):