+++ csfunctions/events/document_released.py
@@ -5,15 +5,15 @@ from pydantic import BaseModel, Field
 from csfunctions.objects import Document, Part
 
 from .base import BaseEvent, EventNames
-from .dialog_data import DocumentReleaseDialogData
+from .dialog_data import DocumentReleasedDialogData
 
 
-class DocumentReleaseData(BaseModel):
+class DocumentReleasedData(BaseModel):
     documents: list[Document] = Field(..., description="List of documents that were released.")
     parts: list[Part] = Field(..., description="List of parts that belong to the released documents")
-    dialog_data: DocumentReleaseDialogData
+    dialog_data: DocumentReleasedDialogData
 
 
-class DocumentReleaseEvent(BaseEvent):
-    name: Literal[EventNames.DOCUMENT_RELEASE] = EventNames.DOCUMENT_RELEASE
-    data: DocumentReleaseData
+class DocumentReleasedEvent(BaseEvent):
+    name: Literal[EventNames.DOCUMENT_RELEASED] = EventNames.DOCUMENT_RELEASED
+    data: DocumentReleasedData